fixed pagination numbering issue
Previously, the page number for the current page would appear as if it were an active link, so it was difficult for the user to tell which page they were currently viewing. Now, the page number for the current page appears as standard bold text, which is a common design pattern in many other pagination implementations.